+++ CHANGELOG.rst
@@ -4,6 +4,9 @@
 - Add `@addable-types` endpoint
   [vangheem]
 
+- Do not allow to add containers inside containers
+  [vangheem]
+
 
 2.2.0 (2017-12-15)
 ------------------